add multithread sampling test
