+++ include/random_walks/random_walks.hpp
@@ -26,5 +26,6 @@
 #include "random_walks/uniform_accelerated_billiard_walk_parallel.hpp"
 #include "random_walks/hamiltonian_monte_carlo_walk.hpp"
 #include "random_walks/langevin_walk.hpp"
+#include "random_walks/multithread_walks.hpp"
 
 #endif // RANDOM_WALKS_RANDOM_WALKS_HPP
