+++ test/CMakeLists.txt
@@ -219,6 +219,15 @@ else ()
   add_test(NAME test_gbaw COMMAND sampling_test -tc=gbaw)
   add_test(NAME test_ghmc COMMAND sampling_test -tc=ghmc)
   
+  add_executable (sampling_multithread_walks sampling_multithread_walks.cpp $<TARGET_OBJECTS:test_main>)
+  add_test(NAME test_multithread_brdhr COMMAND sampling_test -tc=multithread_brdhr)
+  add_test(NAME test_multithread_bcdhr COMMAND sampling_test -tc=multithread_bcdhr)
+  add_test(NAME test_multithread_gcdhr COMMAND sampling_test -tc=multithread_gcdhr)
+  add_test(NAME test_multithread_grdhr COMMAND sampling_test -tc=multithread_grdhr)
+  add_test(NAME test_multithread_biw COMMAND sampling_test -tc=multithread_biw)
+  add_test(NAME test_multithread_cdhr COMMAND sampling_test -tc=multithread_cdhr)
+  add_test(NAME test_multithread_rdhr COMMAND sampling_test -tc=multithread_rdhr)
+  
   add_executable (mmcs_test mmcs_test.cpp $<TARGET_OBJECTS:test_main>)
   add_test(NAME test_mmcs COMMAND mmcs_test -tc=mmcs)
 
@@ -319,6 +328,7 @@ else ()
   add_test(NAME order_polytope_reflection COMMAND order_polytope -tc=reflection)
   add_test(NAME order_polytope_vec_mult COMMAND order_polytope -tc=vec_mult)
 
+  find_package(OpenMP REQUIRED)
 
   TARGET_LINK_LIBRARIES(new_volume_example ${LP_SOLVE} ${MKL_LINK} coverage_config)
 	TARGET_LINK_LIBRARIES(new_volume_example ${LP_SOLVE} coverage_config)
@@ -334,6 +344,7 @@ else ()
   TARGET_LINK_LIBRARIES(new_rounding_test ${LP_SOLVE} ${MKL_LINK} coverage_config)
   TARGET_LINK_LIBRARIES(mcmc_diagnostics_test ${LP_SOLVE} ${MKL_LINK} coverage_config)
   TARGET_LINK_LIBRARIES(sampling_test ${LP_SOLVE} ${MKL_LINK} coverage_config)
+  TARGET_LINK_LIBRARIES(sampling_multithread_walks ${LP_SOLVE} ${MKL_LINK} OpenMP::OpenMP_CXX coverage_config)
   TARGET_LINK_LIBRARIES(mmcs_test ${LP_SOLVE} ${MKL_LINK} coverage_config)
   TARGET_LINK_LIBRARIES(benchmarks_sob ${LP_SOLVE} ${MKL_LINK} coverage_config)
   TARGET_LINK_LIBRARIES(benchmarks_cg ${LP_SOLVE} ${MKL_LINK} coverage_config)
